app/requirements.txt
Django==3.1.5
dj-database-url==0.5.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
whitenoise==5.2.0